Add ScalaTest integration and remove standalone test
- Add ScalaTest dependency and proper test configuration in build.sbt
- Create ProtobufBackportSpec with three test cases for all conversion approaches
- Remove old standalone TestProtobufBackport main class
- Configure JVM options for Java module access in test execution
- Tests now run cleanly with 'sbt test' command
